+++ webapi/repository/student.go
@@ -7,7 +7,7 @@ import (
 )
 
 type StudentRepository interface {
-	Get(context.Context, int64) (*model.Student, error)
+	Get(context.Context, uint32) (*model.Student, error)
 	Set(context.Context, *model.Student) (*model.Student, error)
 	GetAll(context.Context) ([]*model.Student, error)
 	Close() error
